tty bug with multi-line prompts
This effectively reverts 6ec43f691cbe1dbcdddb0aa1c267b338e6b60738
because it was interfering with prompts that spanned more than one
line (by inserting '--More-- + erase' between displaying of prompt and
getting input for the answer).

So we're back to the situation where autodescribe feedback when moving
the cursor will leave text on the second line if it generates text too
wide for one line.  (^R redraws the screen correctly.)