more Guidebook tweaks
Expand the sample configuration file a little and prevent it from
going past the right margin in Guidebook.txt.

Replace all instances of "config file" with "configuration file".

Reformat the "notes" at the end of the table of map symbols.
Unfortunately Guidebook.pdf from Guidebook.ps from Guidebook.mn
puts a page break between the header line "notes" and the two
actual notes.

Value 1 for 'mouse_support' was not just exceeding the margin of
Guidebook.txt but wrapping to the next line.  Shorten it.

Guidebook.tex had a typo "in the foler" (where 'folder' was meant)
and Guidebook.mn didn't have that phrase at all.

Remove a few trailing spaces.