fixes37.0
Remove the entry about creating Guidebook.dat on Unix since it got
moved to 3.6.6.

The orc town raiders were introduced in 3.6.2, not to-be-3.7, so the
fixes entry for unintentionally giving orc zombies and orc mummies
clan names should be in the general fixes section rather than the
exposed by git section.