Sting vs level teleport revisted
This reverses all of fa4538c0bcbeef0960bcd185b8b06c92a9551b6a
except for the fixes37.0 entry and does a better job in a cleaner
fashion.  If Sting is going to start glowing and "you materialize
on a different level" is pending, give the materialize message
before the glowing message.  Otherwise handle both stop-glowing
and/or you-materialize in the normal fashion.