history reconciliation
Eliminate most of the minor differences between dat/history and the
end section of doc/Guidebook.txt which didn't seem to be intentional.
Several commas, a couple of past/present tense discrepancies, and
various clauses or whole sentences which were in one but not the
other.  There are still differences which this doesn't address.

Also a couple of actual changes.  Make the Guidebook refer to itself
as "this document" rather than "this paper".  Change "dozens of
people's work" to "scores of people's work".  Add a new sentence
describing the version numbering scheme used by 3.0, which wasn't
3.0.x yet.