relent on silver bell usage...
Allow a hero in silver-hating form to ring the Bell of Opening when
on the vibrating square so be able to perform the invocation ritual.
But only at that location and only if the stairs aren't there yet.

This is about on par with being able to read the Book of the Dead
while blind although that isn't limited to a specific location.