hints/macOS.2020
Started out removing some trailing spaces and ended up making various
substantive changes.

Don't include tty by default since the sample make command shows how
to enable it and there isn't any easy way to disable it other than
not requesting it to begin with.  (Due to using defined/not-defined
rather than values in the 'if' directives, WANT_WIN_TTY=0 is the same
as WANT_WIN_TTY=1 rather than it's inverse.)

That resulted in all the interfaces starting commented out, so add
some make code to make sure that at least one is enabled.  If none,
it silently enables tty.

The sequence
|CXX=compilerA
|LINK=$(CXX)
|#CXX=compilerB
wouldn't work if 'make' substitutes immediately (I can't recall
offhand whether it does) and the first CXX was commented out in
order to uncomment the second one.  The default CXX value would be
used instead of the #CXX=foo one even if it was uncommented.  Just
move LINK= after it.

Build logic has been tested.  Final install and packaging for
distribution have not (but weren't touched so shouldn't be affected).