interface feedback by #version
When there was only one supported interface included in the program,
feedback of

|Supported windowing system:
|"tty" (traditional text with optional line-drawing)

was missing the intended final period.  When there were more than
one, the clause describing the default could be preceded by a
spurious space.

|Supported windowing systems:
|"tty" (traditional text with optional line-drawing) and "Qt" , with a
|default of "tty".

There was a fixup for " , " but it only worked as intended when
that was on the last line, not when the default's text spanned lines.

This adds description of "Windows GUI" to the mswin entry.  X11 could
have its formal name "The X Windows System" as the description but I
didn't add that since it just seems like extra verbosity.  Apparently
Qt doesn't stand for anything else so still has no extra description.